#49, migration structure added + Room data classes improved

+++ app/src/main/java/com/jumparoundcreations/sugarcounter/data/IEntry.kt
@@ -8,5 +8,5 @@ interface IEntry {
     val currentTimestamp: Long
     val category: String
     val entryType: GramCountMode
-    val gramTotal: Int
+    val gramTotal: Double
 }
